Refactor `wallet` package structure
