+++ pkg/wallet/test_data/testnet.wlt
