+++ pkg/wallet/wallet_test.go
@@ -1,3 +1,6 @@
+// Copyright 2017 The nem-toolchain project authors. All rights reserved.
+// Use of this source code is governed by a MIT license that can be found in the LICENSE file.
+
 package wallet
 
 import (
@@ -5,11 +8,14 @@ import (
 
 	"encoding/hex"
 
+	"io/ioutil"
+	"os"
+
+	"log"
+
 	"github.com/nem-toolchain/nem-toolchain/pkg/core"
 	"github.com/nem-toolchain/nem-toolchain/pkg/keypair"
 	"github.com/stretchr/testify/assert"
-	"fmt"
-	"encoding/base64"
 )
 
 func TestWalletSerialize(t *testing.T) {
@@ -22,23 +28,29 @@ func TestWalletSerialize(t *testing.T) {
 		"M041VlU0SEo2RU5UMjRRVklHQVBNIiwibGFiZWwiOiJQcmltYXJ5IiwibmV0d" +
 		"29yayI6MTA0LCJjaGlsZCI6IjYxM2QwMWNlNjJlNDNjYzViZWE5Mzk1ZTBkOT" +
 		"c5NDJjNDVkNjYxZTA4MTE4NDI0NWRkZWJhZTRlOTc3ZjMzNmYifX19"
-	wlt, _ := Deserialize("./mainnet.wlt")
+	wlt, _ := Deserialize(exp)
 	ser, _ := Serialize(wlt)
-
-	res, _ := base64.StdEncoding.DecodeString(ser)
-	fmt.Println(">> DEBUG  encoded `ser`", string(res))
 
 	assert.Equal(t, ser, exp)
 }
 
 func TestWalletDeserialize(t *testing.T) {
-	wlt, _ := Deserialize("./mainnet.wlt")
+	exp := "eyJwcml2YXRlS2V5IjoiIiwibmFtZSI6Im1haW5uZXQiLCJhY2N" +
+		"vdW50cyI6eyIwIjp7ImJyYWluIjpmYWxzZSwiYWxnbyI6InBhc3M6ZW5j" +
+		"IiwiZW5jcnlwdGVkIjoiZTczZTVlZGFhYzgzOTMzODFhYTFlNWEyN2I3MWJi" +
+		"Y2Q1ODM2ZGY5M2NjZDYwZGMxMTZjOGVjMGI1M2Y0NGQwZTRiZDg0NzJiYWEyM" +
+		"jcyOTcyNjFmNzM4YzY1NjNlNDNkIiwiaXYiOiIxOTBjODVmZjFlNGExNTI2MmZ" +
+		"mOTE3YjgyZDVlOWQ4YyIsImFkZHJlc3MiOiJORExYUzJYSUFWT1BPVkhTVVpJ" +
+		"M041VlU0SEo2RU5UMjRRVklHQVBNIiwibGFiZWwiOiJQcmltYXJ5IiwibmV0d" +
+		"29yayI6MTA0LCJjaGlsZCI6IjYxM2QwMWNlNjJlNDNjYzViZWE5Mzk1ZTBkOT" +
+		"c5NDJjNDVkNjYxZTA4MTE4NDI0NWRkZWJhZTRlOTc3ZjMzNmYifX19"
+	wlt, _ := Deserialize(exp)
 
 	assert.Equal(t, wlt.name, "mainnet")
 	assert.Equal(t, wlt.privateKey, "")
 
 	account := wlt.accounts["0"]
-	exp_address, _ := keypair.FromString("NDLXS2XIAVOPOVHSUZI3N5VU4HJ6ENT24QVIGAPM")
+	exp_address, _ := keypair.ParseAddress("NDLXS2XIAVOPOVHSUZI3N5VU4HJ6ENT24QVIGAPM")
 	assert.Equal(t, account.address, exp_address)
 	assert.Equal(t, account.label, "Primary")
 
@@ -55,14 +67,36 @@ func TestWalletDeserialize(t *testing.T) {
 	assert.Equal(t, account.brain, false)
 }
 
-func TestPrivateKey_EncryptDecrypt(t *testing.T) {
-	pr, _ := hex.DecodeString("2a91e1d5c110a8d0105aad4683f962c2a56663a3cad46666b16d243174673d90")
-	exp, _ := hex.DecodeString("8cd87bc513857a7079d182a6e19b370e907107d97bd3f81a85bcebcc4b5bd3b5")
+func TestReadWalletFile(t *testing.T) {
+	raw := "eyJwcml2YXRlS2V5IjoiIiwibmFtZSI6Im1haW5uZXQiLCJhY2N" +
+		"vdW50cyI6eyIwIjp7ImJyYWluIjpmYWxzZSwiYWxnbyI6InBhc3M6ZW5j" +
+		"IiwiZW5jcnlwdGVkIjoiZTczZTVlZGFhYzgzOTMzODFhYTFlNWEyN2I3MWJi" +
+		"Y2Q1ODM2ZGY5M2NjZDYwZGMxMTZjOGVjMGI1M2Y0NGQwZTRiZDg0NzJiYWEyM" +
+		"jcyOTcyNjFmNzM4YzY1NjNlNDNkIiwiaXYiOiIxOTBjODVmZjFlNGExNTI2MmZ" +
+		"mOTE3YjgyZDVlOWQ4YyIsImFkZHJlc3MiOiJORExYUzJYSUFWT1BPVkhTVVpJ" +
+		"M041VlU0SEo2RU5UMjRRVklHQVBNIiwibGFiZWwiOiJQcmltYXJ5IiwibmV0d" +
+		"29yayI6MTA0LCJjaGlsZCI6IjYxM2QwMWNlNjJlNDNjYzViZWE5Mzk1ZTBkOT" +
+		"c5NDJjNDVkNjYxZTA4MTE4NDI0NWRkZWJhZTRlOTc3ZjMzNmYifX19"
+	exp, _ := Deserialize(raw)
+	wlt, _ := ReadWallet("./test_data/mainnet.wlt")
 
-	pass := derive([]byte("TestTest"))
-	assert.Equal(t, pass, exp)
+	assert.Equal(t, exp, wlt)
+}
 
-	enc, _ := encrypt(pr, pass)
-	dec, _ := decrypt(enc, pass)
-	assert.Equal(t, dec, pr)
+func TestWriteWalletFile(t *testing.T) {
+	exp, _ := ReadWallet("./test_data/mainnet.wlt")
+
+	file, _ := ioutil.TempFile(os.TempDir(), "prefix")
+	defer func() {
+		e := os.Remove(file.Name())
+		if e != nil {
+			log.Fatal(e)
+		}
+	}()
+
+	err := WriteWallet(file.Name(), exp)
+	assert.NoError(t, err)
+
+	wlt, _ := ReadWallet(file.Name())
+	assert.Equal(t, wlt, exp)
 }
